requirements.txt
Flask
psycopg2-binary
scikit-learn
pandas